requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub

pandas
duckdb
requests

dbt-bigquery

dash
plotly
dbt/logs/
logs/
*.log